Conflicts:
	src/visitors/MemberVisitor.cpp

+++ src/firm_interface/FirmInterface.cpp
@@ -396,7 +396,7 @@ std::string FirmInterface::replace_dollar(std::string name)
 {
 	std::string s = name;
 
-	for (int i = 0; i < s.size(); ++i)
+	for (size_t i = 0; i < s.size(); ++i)
 	{
 		if (s[i] == '$')
 			s = s.substr(0, i) + "_C" + s.substr(i + 1, s.size() - i - 1);
